requirements.txt
# Django & REST Framework
Django==4.2.16
djangorestframework==3.14.0
django-redis==5.4.0
django-prometheus==2.3.1
drf-spectacular==0.27.1
django-filter==23.5

# WSGI Server
gunicorn==21.2.0

# Redis
redis==5.0.1

# Utilities
python-dotenv==1.0.0
requests==2.32.5

# Testing
pytest==7.4.3
pytest-django==4.7.0

# Code Quality
flake8==7.0.0
black==23.12.1
